refactored kernel utils
